finished adding new ImageBase class which Trace inherits from
