+++ banzai_nres/tests/test_image.py
@@ -1,7 +1,14 @@
-from banzai_nres.images import NRESImage
-from banzai.tests.utils import FakeContext
-
 import mock
+import numpy as np
+import tempfile
+import os
+from astropy.table import Table
+
+
+from banzai_nres.images import NRESImage, ImageBase
+from banzai.tests.utils import FakeContext, FakeImage
+import banzai_nres.settings as nres_settings
+from banzai import settings
 
 
 @mock.patch('banzai.images.Image._init_instrument_info')
@@ -22,3 +29,44 @@ def test_get_num_lit_fibers(mock_instrument):
     image = NRESImage(runtime_context=FakeContext(),
                       header={'OBJECTS': 'none&tung&none'})
     assert image.num_lit_fibers() == 1
+
+
+class TestImageBase:
+    def test_load_and_write(self):
+        name = 'trace'
+        image = ImageBase(data=Table({'id': [1], 'centers': [np.arange(3)]}), table_name=name)
+        runtime_context = FakeContext()
+        with tempfile.TemporaryDirectory() as tmp_directory_name:
+            runtime_context.fpack = False
+            path = os.path.join(tmp_directory_name, 'test_trace_table.fits')
+            image.filepath = path
+            image.header = {'bla': 1}
+            image.write(runtime_context, update_db=False)
+            loaded_image = ImageBase.load(path=path, extension_name=name)
+            assert np.allclose(loaded_image.data['centers'][0], image.data['centers'][0])
+            assert np.allclose(loaded_image.data['id'][0], image.data['id'][0])
+            assert np.isclose(loaded_image.header['bla'], 1)
+
+    def test_write_gets_correct_filename(self):
+        name = 'trace'
+        trace = ImageBase(data=Table({'id': [1], 'centers': [np.arange(3)]}), table_name=name)
+        runtime_context = FakeContext()
+        with tempfile.TemporaryDirectory() as tmp_directory_name:
+            for fpack, extension in zip([True, False], ['.fz', 'its']):
+                runtime_context.fpack = fpack
+                path = os.path.join(tmp_directory_name, 'test_trace_table.fits')
+                trace.filepath = path
+                trace.header = {'bla': 1}
+                trace._update_filepath(runtime_context)
+                assert trace.filepath[-3:] == extension
+
+    def test_instantiates_db_context_from_image(self):
+        image = FakeImage()
+        possible_attributes = ['dateobs', 'datecreated', 'instrument', 'is_master', 'is_bad']
+        counter = np.arange(len(possible_attributes))
+        for attribute, i in zip(possible_attributes, counter):
+            setattr(image, attribute, str(i))
+        imagebase = ImageBase(data={'id': [1], 'centers': [np.arange(3)]}, image=image, obstype='TRACE')
+        for attribute in possible_attributes:
+            assert getattr(imagebase, attribute) == getattr(image, attribute)
+        assert imagebase.attributes == settings.CALIBRATION_SET_CRITERIA.get('TRACE', {})
